Some refacturing and two big bugs fixed
1. I changed a lot of variables names.

2. I renamed "InitializeArrayOfMines" method to
"InitializeNewGameField", and I maked total refactoring.

3. I renamed "CountOpen" method to "CountOpenCells" and renamed some
variables inside.

4. I renamed "DisplayTop" method to "DisplayScoreBoard" and make some
refactoring(more about when I explane about the bugs).

Bug 1:
The top command didn't work! So i changed the loop in DisplayScoreBoard
- his lenght was "playerScoreCounter % 6" and I maked to
"playerScore.Length"!
After this i maked playerScoreCounter  to increases his value with 1
after every move. And finaly when the game is over a changed the way for
saving the score and name of the player. I sort "playerScore" array and
check if the new score is bigger than some of the old ones. And if it is
true I put the new score at his place and new player name in the place
of the name in the same index in "playerName" array.

Bug 2:
After the player loose the game and enter his name, the game starts
again but on the old game field. And after the last game is over the
field is open so the player always have wrong input. So after each game
I initialise and display new field.