Added new GameField method
Added a new GameField method to be used when we separate the Telerik.cs
file in a more OOP approach.

1. Added arrayOfMines, randomNumbers, gameField as properties. Changed
gameField to Field.

2. Added a contructor that initializes the three arrays with the same
size.

3. All Display- methods are added. All are changed so that they return a
string.
DisplayGameField is renamed ToString().

4. CountSurroundingMines, InitializeNewGameField are added as public
methods. IsPositionValid, CountOpenCells, GetSingleFieldRowString,
ClearTheField, InitialiseTheMinesOnTheField, IsFoundInRandomNumbers are
added as private nonstatic methods.

5. randomNumbers is used in only two methods, so now it is initialized
in InitialiseTheMinesOnTheField and passed to IsFoundInRandomNumbers as
a
parameter.